code refactor/re-arainge files
